# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: axe API
  version: "1.0.0"
  description: |
    Production-grade Go REST API built with axe framework.
    Clean Architecture — Chi router, Ent ORM, JWT auth, Redis cache, Prometheus metrics.
  contact:
    name: axe team
    url: https://github.com/henrynguyen6677/axe
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development

# ── Security ──────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token. Obtain via POST /api/v1/auth/login.
        Include as: `Authorization: Bearer <token>`
  schemas:
...
    UserRole:
      type: string
      enum: [user, admin]
    User:
      type: object
      required: [id, email, name, role, active, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        email:
          type: string
          format: email
          example: "alice@example.com"
        name:
          type: string
          example: "Alice Smith"
        role:
          $ref: "#/components/schemas/UserRole"
        active:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateUserRequest:
      type: object
      required: [email, name, password]
      properties:
        email:
          type: string
          format: email
          example: "alice@example.com"
        name:
          type: string
          minLength: 1
          example: "Alice Smith"
        password:
          type: string
          minLength: 8
          format: password
          example: "super_secret_123"
        role:
          $ref: "#/components/schemas/UserRole"
...
    ListUsersResponse:
      type: object
      required: [data, total, limit, offset]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
          example: 42
        limit:
          type: integer
          example: 20
        offset:
          type: integer
          example: 0

    # ── Auth ────────────────────────────────────────────────────────────────
...
paths:
...
  /api/v1/users:
    get:
      summary: List users
      operationId: listUsers
      tags: [users]
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        "200":
          description: Paginated list of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListUsersResponse"
        "401":
          description: Not authenticated

    post:
      summary: Create user
      operationId: createUser
      tags: [users]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Validation error
        "409":
          description: Email already exists
...